Cleaned up SQL query and added MySQLi escape strings.
I used the ` backtick around the column names to clean up the query. I
changed the database table from documents to ruskin.documents to be more
specific that we're adding this to a database named ruskin. I also used
the real_escape_string() function to add character escapes to all of the
variables we're inserting into the database. This is standard practice
to ALWAYS escape your variables whenever they're used, so that #1 your
query doesn't break if one of your variables has a quote or double
quote, and #2 for security reasons so that a malicious user can't use a
malformed query to hijack our MySQL server.