URL is now being added to the database (WORK IN PROGRESS)
The URL field is calculated and is now being added to the database. I
added SOME of the hardcoded links that we need, such as when title links
to witnesses, etc. The rest of these hardcoded links still need to be
finished though.

I also added some code to visit the resulting URL and check if the file
exists on the website. If the XML file is not being used on the website,
then it is skipped from being added to the database.