redesign with checkboxes in banner
Design wise this saves a lot of wasted space and I believe is much more
pleasing to the eye.
But, I have to fix the fact that its not centered correctly. Garth if ya
have a chance, take a look and tell me what you see. Doesn't have the
educational part yet. I need to fix that centering..