Missing files are now being indexed
About a third of our XML files were not being indexed because their HTML
equivalent could not be located on the site. After meeting with the
client, the proper locations were determined. These documents are now
being indexed.

I also added counters to show the number of successful keywords, and the
files that were skipped because they were not needed.

In addition, I cleaned up our code.